+++ CHANGES.md
@@ -6,6 +6,8 @@
 - For an enum `Foo(A,B,C)`, the syntax `f!` now desugars to `f.A!` using the
   first tag defined in the enum.
 - Error messages are more helpful for `foo.Whatever!` enum field accessing.
+- Simplified logic for enums so there is less difference between enums that
+  have tags with member fields and those without.
 
 ## v2025-11-30